Asyncify TestRetryAfter and TestIPv6 PoolManager tests
